refactor: move config runtime resolution out of config
